Added navigation controller with the current view controller as a root
Notice that I've given a title to the initial view controller to be "Initial View Controller". This is automatically marked up in the storyboard xml to be a commented section. You can notice that there is also now a navigationItem added to the initial view controller with the title "Initial View Controller". You can also see that there is a navigation controller added with a "connections" tag and a segue to the view controller. You can tell, because they both share the id of "BYZ-38-t0r". This segue is declaring that the relationship between the two controllers is that the view controller is a "rootTableView" of the navigation controller.