Added navigation to another view controller
I added a "Next" barButtonItem to the navigationItem, and created a segue that takes the user to the "Next View Controller". When using MVVMC, we wouldn't use this segue, we would simply hook this button up to an IBAction that calls the coordinator and fetches the next view controller from the segue as I'll display in the next commit.